refactor: added default implementation for observe_update and updated atom & atom:undo & reaction & access_state
