+++ src/reactive_state_access/atom.rs
@@ -258,18 +258,6 @@ where
     fn id(&self) -> StorageKey {
         self.id
     }
-
-    #[topo::nested]
-    fn observe_update(&self) -> (Option<T>, T)
-    where
-        T: 'static + Clone,
-    {
-        let previous_value_access = crate::hooks_state_functions::use_state(|| None);
-        let opt_previous_value = previous_value_access.get();
-        let new_value = self.get();
-        previous_value_access.set(Some(new_value.clone()));
-        (opt_previous_value, new_value)
-    }
 }
 // The below is broke as need None if no prior state
 impl<T> ObserveChangeReactiveState<T> for Atom<T>
@@ -423,7 +411,6 @@ where
     }
 }
 
-use crate::reactive_state_access::state_access::CloneState;
 use crate::reactive_state_access::{CloneReactiveState, ObserveChangeReactiveState};
 use crate::{
     clone_reactive_state_with_id, reactive_state_exists_for_id,
