+++ src/reactive_state_access/atom_undo.rs
@@ -1,4 +1,3 @@
-use crate::reactive_state_access::state_access::CloneState;
 use crate::reactive_state_access::CloneReactiveState;
 use crate::{
     clone_reactive_state_with_id, reactive_state_exists_for_id,
@@ -93,18 +92,6 @@ where
 {
     fn id(&self) -> StorageKey {
         self.id
-    }
-
-    #[topo::nested]
-    fn observe_update(&self) -> (Option<T>, T)
-    where
-        T: 'static + Clone,
-    {
-        let previous_value_access = crate::hooks_state_functions::use_state(|| None);
-        let opt_previous_value = previous_value_access.get();
-        let new_value = self.get();
-        previous_value_access.set(Some(new_value.clone()));
-        (opt_previous_value, new_value)
     }
 }
