+++ src/reactive_state_access/observable.rs
@@ -1,3 +1,5 @@
+use crate::reactive_state_access::state_access::CloneState;
+use crate::reactive_state_access::CloneReactiveState;
 use crate::reactive_state_functions::STORE;
 use crate::store::StorageKey;
 use crate::{clone_reactive_state_with_id, read_reactive_state_with_id, ReactiveContext};
@@ -29,9 +31,18 @@ where
 
         clone_reactive_state_with_id::<T>(self.id()).unwrap()
     }
+    #[topo::nested]
     fn observe_update(&self) -> (Option<T>, T)
     where
-        T: Clone + 'static;
+        T: 'static + Clone,
+        Self: CloneReactiveState<T>,
+    {
+        let previous_value_access = crate::hooks_state_functions::use_state(|| None);
+        let opt_previous_value = previous_value_access.get();
+        let new_value = self.get();
+        previous_value_access.set(Some(new_value.clone()));
+        (opt_previous_value, new_value)
+    }
     fn observe_with<F: FnOnce(&T) -> R, R>(&self, func: F) -> R {
         if let Ok(context) = illicit::get::<RefCell<ReactiveContext>>() {
             context
